+++ tests/unit/Command/DockerManageCommandTest.php
@@ -3,7 +3,7 @@
  * Linux for PHP/Linux for Composer
  *
  * Copyright 2017 - 2020 Foreach Code Factory <lfphp@asclinux.net>
- * Version 2.0.2
+ * Version 2.0.3
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -43,7 +43,7 @@ class DockerManageCommandTest extends KernelTestCase
 
     protected $progressBarMock;
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
         if (!defined('PHARFILENAME')) {
             define(
@@ -92,7 +92,7 @@ class DockerManageCommandTest extends KernelTestCase
         }
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         \Mockery::close();