+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 2.0.3 (2020-04-09)
+
+- Adds the shared Docker volume size feature for the LfPHP Cloud.
+
 ## 2.0.2 (2020-03-19)
 
 - Fixes an issue when changing from a Dockerfile to Linux for Composer in order to start containers.