+++ docs/whats_new.rst
@@ -3,6 +3,12 @@
 What's New
 ==========
 
+What's New in Version 2.0.3 (2020-04-09)
+----------------------------------------
+
+- Adds the shared Docker volume size feature for the LfPHP Cloud.
+
+
 What's New in Version 2.0.2 (2020-03-19)
 ----------------------------------------